+++ index.js
@@ -12,8 +12,10 @@ module.exports = function(path) {
       ...require('./Sequential/ff'),
       ...require('./Sequential/registers'),
       ...require('./Sequential/counters'),
-      ...require('./Sequential/memory')
+      ...require('./Sequential/memory'),
+      ...require('./Sequential/cpu')
     },
+    'CPU': require('./Sequential/cpu'),
     'IO': require('./Utility/ioManager'),
     'Base': require('./Utility/new')
   } [ path ]